feat(core): rename base -> bases
