+++ packages/core/src/bases/typedef.ts
