+++ packages/core/src/index.ts
@@ -18,7 +18,7 @@ import { installShims } from './common/shims';
 
 installShims();
 
-export * from './base';
+export * from './bases';
 export { dedupe, dedupeBy, groupBy, makeArray, remove, rotate } from './common/array';
 export * from './common/async';
 export { isBooleanString } from './common/boolean';
